server/requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-dotenv
pymongo
pyjwt
bcrypt
